Restructure repository around 'optimake' package
