+++ examples/bzipped_pymatgen/.testing/first_entry.json
