+++ examples/bzipped_pymatgen/optimade.yaml
