+++ examples/bzipped_pymatgen/part_1.json.bz2
