+++ examples/folder_of_cifs/optimade.yaml
