+++ examples/xyz_files_no_compression/.testing/first_entry.json
