+++ examples/xyz_files_no_compression/C_1.xyz
