+++ examples/xyz_files_no_compression/H_1.xyz
