+++ examples/xyz_files_no_compression/H_2.xyz
