+++ examples/xyz_files_no_compression/optimade.yaml
