--- src/mc_optimade/mc_optimade/__init__.py
@@ -1,4 +1,0 @@
-from mc_optimade.config import Config
-from mc_optimade.convert import convert_archive
-
-__all__ = ("Config", "convert_archive")
