+++ src/optimake/__init__.py
@@ -1,0 +1,4 @@
+from optimake.config import Config
+from optimake.convert import convert_archive
+
+__all__ = ("Config", "convert_archive")
