+++ src/optimake/archive/cli.py
@@ -1,5 +1,5 @@
 import click
-from scan_records import scan_records
+from .scan_records import scan_records
 
 
 @click.group()
