+++ src/optimake/parsers.py
@@ -10,7 +10,7 @@ from optimade.adapters import Structure
 from optimade.models import EntryResource
 from pymatgen.entries.computed_entries import ComputedStructureEntry
 
-from mc_optimade.config import PropertyDefinition
+from optimake.config import PropertyDefinition
 
 
 def pybtex_to_optimade(bib_entry: Any, properties=None) -> EntryResource:
