+++ tests/test_yaml.py
@@ -2,7 +2,7 @@ from pathlib import Path
 
 import pytest
 
-from mc_optimade.config import Config
+from optimake.config import Config
 
 EXAMPLE_YAMLS = (Path(__file__).parent.parent / "examples").glob("*/optimade.yaml")
